Properly read ASCII packet bytes
Previously, the output from the backend was buffered and the contents
of the buffer were filled via `BufReader::fill_buf()`. However, this
function only reads from the reader if its buffer is empty. So, if the
buffer ever contained a partial packet, more data would never be read
and an error would be returned, despite more data likely being available
to be read.

This change removes the buffer and reads from the backend until a
properly framed packet is found.