Refactor the `ascii::Target` API
The previous API was a mix of constructors and getters. However,
the constructors looked like getters and the getters didn't follow
the Rust API guidelines (they started with `get_`). This change makes
the constructors read more clearly as constructors and makes the getters
follow the API guidelines. It also expands the documentation to avoid
confusion.