Rename `Port::on_*` and related methods
`on_*` sets a value but doesn't follow the naming convention used for
other setters. Additionally, the `clear_on_*` methods read oddly making
their meaning confusing. This change renames the `on_*` methods to
`set_*_handler` and `clear_on_*` to `clear_*_handler`, which is much
clearer and consistent with the rest of the API.