Remove unnecessary docs
Previously the *Handlers types were exposed in the main `ascii` module
with all the other types. They were an implementation detail and
documented as such, so people could learn the API from the docs more
easily. However, now they are exposed in a submodule, so there is no
need to document them as an implementation detail.