Delete binary compiler-output tests
The binary code hasn't changed in years yet the tests fail every year
because the diagnostic output of the compiler isn't stable. Additionally
now that unstable code runs against the nightly compiler failures are
often guaranteed as the tests cannot adhere to the outputs of both
stable and nightly branches of the compiler.

Since the risk is minimal and the overhead is high, this change removes
the binary compile time tests.